services/api/pyproject.toml
[tool.poetry]
name = "api"
version = "0.1.0"
description = ""
authors = ["EnriqueColpo <EnriqueColpo@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.111.0"
uvicorn = "^0.30.1"
httpx = "^0.27.0"
mangum = "^0.17.0"
boto3 = "^1.34.127"
ipdb = "^0.13.13"
jwt = "^1.3.1"
pyjwt = "^2.8.0"
pydantic-settings = "^2.3.3"


[tool.poetry.group.dev.dependencies]
pytest = "^8.2.2"
pytest-cov = "^5.0.0"
black = "^24.4.2"
isort = "^5.13.2"
flake8 = "^7.1.0"
bandit = "^1.7.9"
moto = "^5.0.9"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
